--- src/main/java/com/kwakmunsu/vote/response/exception/StatusItem.java
@@ -1,14 +1,0 @@
-package com.kwakmunsu.vote.response.exception;
-
-import org.springframework.http.HttpStatus;
-
-public class StatusItem {
-    public static final int OK = HttpStatus.OK.value();
-    public static final int CREATED = HttpStatus.CREATED.value();
-    public static final int NO_CONTENT = HttpStatus.NO_CONTENT.value();
-    public static final int BAD_REQUEST =  HttpStatus.BAD_REQUEST.value();
-    public static final int UNAUTHORIZED = HttpStatus.UNAUTHORIZED.value();
-    public static final int FORBIDDEN = HttpStatus.FORBIDDEN.value();
-    public static final int NOT_FOUND = HttpStatus.NOT_FOUND.value();
-    public static final int INTERNAL_SERVER_ERROR = HttpStatus.INTERNAL_SERVER_ERROR.value();
-}
